requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
Django==5.0.2
djangorestframework==3.14.0
iniconfig==2.0.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pytest==8.0.1
pytest-django==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
ruff==0.2.2
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
